mcp/pyproject.toml
[project]
name = "datatist-mcp-server"
version = "0.1.0"
description = "Sample MCP Server for Data Analytics Platform"
authors = [
    {name = "datatist", email = "xin.zhao@datatist.com"}
]
requires-python = ">=3.12"
dependencies = [
    "cryptography>=46.0.3",
    "fastmcp>=0.1.0",
    "httpx>=0.28.1",
    "pycryptodome>=3.23.0",
    "pytest>=9.0.2",
    "uvicorn>=0.20.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]

[project.scripts]
wolf-backend-mcp = "wolf_backend_mcp:main"
